requirements.txt
black==24.1.1
flake8==7.0.0
isort==5.13.2
pillow==12.2.0
